Particle system tweaks
Changed emit to its own pipeline instead of branching inside simulate.
This allows using indirect dispatch to start simulate with a dispatch
size matching the amount of alive particles, instead of always using the
maximum capacity of the system.
Added a compaction step that moves all alive particles to the beginning
of the buffer. Emit adds particles to the end of the buffer using atomic
ops.