Add subgroup operations in particle simulate shaders
Added subgroup operations to reduce contention on atomic operations
performed on the global resources. This should theoretically increase
performance, but this is hard to measure currently, as the framerate is
so high that the number of particles added or removed during any given
frame is very low. Enabling vsync seems to lower performance
so much that measuring with vsync is not meaningful.