Added building acceleration structures for particles
Indirect builds are not supported on the tested hardware (RTX 4070),
so particles are built using a build size set to the maximum particle
capacity and dead particles are set to have degenerate AABBs.
Building a BLAS in this way with each alive particle constituting one
AABB was quite slow, causing the simulate step to take 1.7ms instead of
the previous 0.1 ms. This might be more feasible with lower particle
counts.
Tested acceleration structure correctness by tracing particles as
spheres.