Tested raytracing shadows for particles
This has catastrophic performance for the this use case, as the shadows
would have to be traced per fragment and overdraw caused by overlapping
fragments causes the number of rays evaluated per pixel to blow up.

Also tested putting all particles in a single BLAS vs using a single
BLAS that is instanced for each particle in the TLAS. No significant
difference for small particle counts, instanced TLAS was significantly
slower for higher particle counts with lots of 'dead' particles
(degenerate instances).